apps/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "FastAPI application for Recruity AI"
authors = ["Recruity AI <dev@recruity.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
